release: staging → main (3 XL tickets — #295 #293 #286) (#383)
* feat(opt): seamless bootstrap, offline toasts, latency health, perf spot-check (#336 #101 #313 #263) (#372)

- AppRoot: defer dismissStaticLoadingShell() until cached data loads; add
  online/offline listeners with toast feedback (#336 #101).
- Entry: defer LandingModal until appBootstrapStore.phase === 'ready' (#336).
- middleware.ts: record all request latencies via latency-tracker (#313).
- /api/health: JSON endpoint with p50/p95/p99 per route and overall (#313).
- scripts/performance-spot-check.sh: add total JS payload, gzip ratio, and
  /api/health probe (#263).
- latency-tracker.test.ts: unit tests for percentile and filtering.

* Batch: AMIS upsert, offline categories, SQLite cleanup, a11y (#287 #294 #300 #337 #269) (#374)

* feat: batch five issues — AMIS upsert, offline categories, a11y (#287 #294 #300 #337 #269)

Upsert AMIS imports with change summaries and room alias resolution, split offline downloads into map vs directory, remove unused better-sqlite3, and apply a small accessibility pass on search and term controls.

* feat(batch): 10 small tickets — route zoom, PWA install, class search, schedule tokens, OG meta, a11y (#94 #213 #246 #240 #241 #128 #268 #269 #270 #237)

- Map.svelte: fitBounds on 2-point walking routes (#94).
- LandingModal: PWA install CTA via beforeinstallprompt (#213).
- ClassQuery: retry button on class fetch error (#246).
- schedule-renderer.ts: read design tokens from CSS custom properties (#240);
  responsive canvas sizing with DPR scaling (#241).
- Entry.svelte: prefers-reduced-motion support (#269).
- og-meta.test.ts: unit tests for OG image URL and absoluteUrl (#268).
- version.test.ts: existing tests verify footer version label (#270).

* chore: format batch files and fix entity action chip overflow (#378)

Prettier pass on #374 files so CI verify passes; includes action chip
label overflow fix from #377.

* Batch: keyboard shortcuts, hover previews, overlay stacking, offline schedules (#380)

* Add keyboard shortcuts, hover previews, overlay stacking, and offline schedules

Ships v1 global shortcuts (/, Ctrl+K, ?, T) with help chip, map pin hover tooltips, ephemeral overlay dismissal when blocking UI opens, and offline class schedule downloads (#269, #288, #302, #337).

* refactor(stores): split store.svelte.ts into domain modules (#295)

- Create src/lib/stores/ with domain files:
  - ui-stores.svelte.ts (Modal, Query, Toast, MainControls, FloatingControlPanel)
  - map-stores.svelte.ts (Map, MapView, MapTools, Terrain, Building3D)
  - editor-stores.svelte.ts (EditorChrome, MapEdit, MapProposal, AdditionProposal, EventPlacement)
  - sync-stores.svelte.ts (AppBootstrap, SyncToast, Offline)
  - data-stores.svelte.ts (Term, RoomClasses)
- Coupled stores (Location, Jeepney, AdminAuth, Proposals, ScheduleRoute) stay in index.ts
- store.svelte.ts is now a thin barrel: export * from './stores/index.js'
- All @lib/store.svelte imports work unchanged

* infra(auth): consolidate editor auth on Supabase (#293)

- Add supabase_user_id uuid column to admin_users (drizzle schema + migration 0015)
- Add getAdminUserBySupabaseId() to admin-user-service
- Update POST /api/admin/auth to try Supabase signInWithPassword first,
  then fall back to bcrypt, then legacy password
- Update middleware to resolve editorUser from HMAC cookie OR Supabase session
- AdminAuthStore unchanged — server-side handles Supabase integration

* feat(data): AMIS JSON import script for Supabase classes (#286)

- scripts/import-amis-classes.ts: reads AMIS JSON, maps facility_id → room_id,
  resolves term_id, replaces existing term classes, reports unmatched facilities
- Add npm script: bun run import:amis-classes -- --file … --term-id …
- Prettier pass on store domain modules and middleware
