+++ src/components/svelte/Map.svelte
@@ -35,7 +35,7 @@
   import { observeBlockHeight } from "@lib/layout-css-vars";
   import * as mapGl from "maplibre-gl";
   import type { FeatureCollection, LineString } from "geojson";
-  import type { EventData } from "@lib/types";
+  import type { BuildingData, EventData } from "@lib/types";
   import {
     JEEPNEY_ROUTES,
     type JeepneyRoute,
@@ -76,6 +76,11 @@
     ClientEventConflictError,
     editErrorMessage,
   } from "@lib/map-edit/errors";
+  import {
+    buildingPreviewFromRow,
+    entityHoverPreviewStore,
+    eventPreviewFromRow,
+  } from "@lib/entity-hover-preview.svelte";
   import { patchEventLocations, patchPosition } from "@lib/map-edit/patch-api";
   import type {
     EditableCoords,
@@ -756,6 +761,46 @@
   function handleEditablePinLeave(key: string) {
     if (!canDragPin(key)) return;
     if (hoveredEditKey === key) hoveredEditKey = null;
+  }
+
+  function shouldShowEntityHoverPreview() {
+    return (
+      !mapEditStore.enabled &&
+      !mapProposalStore.enabled &&
+      !eventPlacementStore.active
+    );
+  }
+
+  function handleBuildingPinPointerEnter(
+    building: BuildingData,
+    editKey: string,
+    event: PointerEvent,
+  ) {
+    handleEditablePinEnter(editKey);
+    if (!shouldShowEntityHoverPreview() || canDragPin(editKey)) return;
+    entityHoverPreviewStore.show(buildingPreviewFromRow(building), {
+      x: event.clientX,
+      y: event.clientY,
+    });
+  }
+
+  function handleBuildingPinPointerLeave(editKey: string) {
+    handleEditablePinLeave(editKey);
+    if (!shouldShowEntityHoverPreview() || canDragPin(editKey)) return;
+    entityHoverPreviewStore.scheduleHide();
+  }
+
+  function handleEventPinPointerEnter(event: EventData, pointer: PointerEvent) {
+    if (!shouldShowEntityHoverPreview()) return;
+    entityHoverPreviewStore.show(eventPreviewFromRow(event), {
+      x: pointer.clientX,
+      y: pointer.clientY,
+    });
+  }
+
+  function handleEventPinPointerLeave() {
+    if (!shouldShowEntityHoverPreview()) return;
+    entityHoverPreviewStore.scheduleHide();
   }
 
   function beginMarkerDrag(key: string) {
@@ -2455,6 +2500,9 @@
                       entry.event.occurrenceStartsAt,
                     )}`}
                     onclick={() => handleEventMarkerClick(entry.event)}
+                    onpointerenter={(event) =>
+                      handleEventPinPointerEnter(entry.event, event)}
+                    onpointerleave={handleEventPinPointerLeave}
                   />
                 {/if}
               {:else}
@@ -2480,6 +2528,9 @@
                     title={`${group.entries.length} events at ${group.label}`}
                     ariaLabel={`${isExpanded ? "Collapse" : "Expand"} ${group.entries.length} events at ${group.label}`}
                     onclick={() => toggleEventMarkerGroup(group.key)}
+                    onpointerenter={(event) =>
+                      handleEventPinPointerEnter(primaryEntry.event, event)}
+                    onpointerleave={handleEventPinPointerLeave}
                   />
                   {#if isExpanded}
                     <div
@@ -2604,8 +2655,9 @@
                     : "idle"}
               title={building.buildingName}
               labelVisible={zoomLevel >= 17 || hoveredEditKey === editKey}
-              onpointerenter={() => handleEditablePinEnter(editKey)}
-              onpointerleave={() => handleEditablePinLeave(editKey)}
+              onpointerenter={(event) =>
+                handleBuildingPinPointerEnter(building, editKey, event)}
+              onpointerleave={() => handleBuildingPinPointerLeave(editKey)}
             >
               <University size="20" />
             </MapEntityPin>
@@ -2677,7 +2729,7 @@
                     : "idle"}
               title={dorm.dormName}
               labelVisible={zoomLevel >= 17 || hoveredEditKey === editKey}
-              onpointerenter={() => handleEditablePinEnter(editKey)}
+              onpointerenter={(event) => handleEditablePinEnter(editKey)}
               onpointerleave={() => handleEditablePinLeave(editKey)}
             >
               <House size="18" />
