+++ src/components/svelte/TermSelector.svelte
@@ -7,6 +7,7 @@
   import { trapFocus } from "@lib/focus-trap";
   import { portal } from "@lib/portal";
   import { mapToolsStore, termStore } from "@lib/store.svelte";
+  import { registerEphemeralOverlayDismisser } from "@lib/overlay-stack";
   import type { TermWithCount } from "@lib/types";
   import "./map-chrome/map-chrome.css";
 
@@ -23,6 +24,17 @@
 
   onMount(() => {
     termStore.init();
+    const onOpenRequest = () => {
+      if (!open) toggleOpen();
+    };
+    window.addEventListener("room-tba:open-term-picker", onOpenRequest);
+    const unregisterDismiss = registerEphemeralOverlayDismisser(() => {
+      open = false;
+    });
+    return () => {
+      window.removeEventListener("room-tba:open-term-picker", onOpenRequest);
+      unregisterDismiss();
+    };
   });
 
   const active = $derived(termStore.activeTerm);
