+++ src/lib/search-focus.ts
@@ -1,0 +1,12 @@
+let focusSearchInput: (() => void) | null = null;
+
+export function registerSearchFocus(handler: () => void) {
+  focusSearchInput = handler;
+  return () => {
+    if (focusSearchInput === handler) focusSearchInput = null;
+  };
+}
+
+export function focusSearch() {
+  focusSearchInput?.();
+}
