+++ src/pages/api/admin/auth.ts
@@ -15,7 +15,9 @@ import {
 import {
   authenticateAdminUser,
   authenticateLegacyAdminPassword,
+  getAdminUserBySupabaseId,
 } from "@lib/services/admin-user-service";
+import { createServerSupabaseClient } from "@lib/supabase/server";
 
 export const prerender = false;
 
@@ -41,7 +43,7 @@ export const GET: APIRoute = async ({ cookies }) => {
   );
 };
 
-export const POST: APIRoute = async ({ request }) => {
+export const POST: APIRoute = async ({ request, cookies }) => {
   const ip = clientIp(request);
   const ipRate = checkRateLimit(
     `admin-login:ip:${ip}`,
@@ -73,7 +75,28 @@ export const POST: APIRoute = async ({ request }) => {
     return json({ error: "Password is required" }, 400);
   }
 
-  let user = username ? await authenticateAdminUser(username, password) : null;
+  let user: Awaited<ReturnType<typeof authenticateAdminUser>> = null;
+
+  // Try Supabase Auth when configured (#293)
+  try {
+    const supabase = createServerSupabaseClient({
+      request,
+      cookies,
+    });
+    const { data, error } = await supabase.auth.signInWithPassword({
+      email: username,
+      password,
+    });
+    if (data.user && !error) {
+      user = await getAdminUserBySupabaseId(data.user.id);
+    }
+  } catch {
+    // Supabase not configured or unavailable → fall through to bcrypt
+  }
+
+  if (!user) {
+    user = username ? await authenticateAdminUser(username, password) : null;
+  }
 
   if (!user && !username) {
     user = await authenticateLegacyAdminPassword(password);
